separate slow-running integration tests into their own file with a build tag of "integration"
