+++ pkg/storage/mysql/mysql_test.go
@@ -1,3 +1,5 @@
+//go:build integration
+
 package mysql
 
 import (
