+++ pkg/storage/postgres/postgres_test.go
@@ -1,3 +1,5 @@
+//go:build integration
+
 package postgres
 
 import (
