+++ pkg/storage/sqlite/sqlite_test.go
@@ -1,3 +1,5 @@
+//go:build integration
+
 package sqlite
 
 import (
